+++ types/src/timestamp.rs
@@ -372,6 +372,34 @@ impl From<Duration> for TimeDiff {
     }
 }
 
+/// A module for the `[serde(with = serde_option_time_diff)]` attribute, to serialize and
+/// deserialize `Option<TimeDiff>` treating `None` as 0.
+pub mod serde_option_time_diff {
+    use super::*;
+
+    /// Serializes an `Option<TimeDiff>`, using `0` if the value is `None`.
+    pub fn serialize<S: Serializer>(
+        maybe_td: &Option<TimeDiff>,
+        serializer: S,
+    ) -> Result<S::Ok, S::Error> {
+        maybe_td
+            .unwrap_or_else(|| TimeDiff::from(0))
+            .serialize(serializer)
+    }
+
+    /// Deserializes an `Option<TimeDiff>`, returning `None` if the value is `0`.
+    pub fn deserialize<'de, D: Deserializer<'de>>(
+        deserializer: D,
+    ) -> Result<Option<TimeDiff>, D::Error> {
+        let td = TimeDiff::deserialize(deserializer)?;
+        if td.0 == 0 {
+            Ok(None)
+        } else {
+            Ok(Some(td))
+        }
+    }
+}
+
 impl From<u64> for TimeDiff {
     fn from(duration: u64) -> TimeDiff {
         TimeDiff(duration)